[ThumbScrollView, DetailStackView] 추가
